% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{charVecToSingleLength}
\alias{charVecToSingleLength}
\title{Convert any will character vector to a single length character vector}
\usage{
charVecToSingleLength(x)
}
\arguments{
\item{x}{a \code{character}}
}
\value{
x a single-length character vector Non-NA
}
\description{
Convert any will character vector to a single length character vector
}
\examples{
tinytest2JUnit:::charVecToSingleLength(c("Hello", "World")) # -> "HelloWorld"
tinytest2JUnit:::charVecToSingleLength(c("Hello", NA_character_)) # -> "HelloNA"
tinytest2JUnit:::charVecToSingleLength(character(0L)) # -> ""
}
