% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructJUnitTag.R
\name{classnameTestcase}
\alias{classnameTestcase}
\title{Helper function specifying the 'classname' attribute of the testcase tag.}
\usage{
classnameTestcase(tinytest)
}
\arguments{
\item{tinytest}{a \code{tinytest}-object representing an individual test case.}
}
\value{
\code{character(1)} being the 'classname'
}
\description{
Helper function specifying the 'classname' attribute of the testcase tag.
Currently equal to the fileName. The classname is already xml escaped.
}
