% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructJUnitTag.R
\name{constructFailureDescription}
\alias{constructFailureDescription}
\title{Helper function generating the body of a failure description tag!}
\usage{
constructFailureDescription(tinytest)
}
\arguments{
\item{tinytest}{A \code{tinytest} objec that is considered failed!.}
}
\value{
\code{character(1)} being the failure tag description body. This string is already propery
xml escaped.
}
\description{
Helper function generating the body of a failure description tag! Attempts to mimic the print
behaviour of a tinytest object.
}
