% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructJUnitTag.R
\name{constructTestcaseTag}
\alias{constructTestcaseTag}
\title{Construct JUnit testcase tag}
\usage{
constructTestcaseTag(tinytest)
}
\arguments{
\item{tinytest}{a \code{\link[tinytest:tinytest]{tinytest::tinytest()}}-object representing an individual test case.}
}
\value{
\code{XMLtag}: with tag-name = "tinytest" and contains all the tests results of the file.
}
\description{
Construct JUnit testcase tag based on a single tinytest results.
}
\author{
Lennart Tuijnder
}
