% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructJUnitTag.R
\name{constructTeststuitesTag}
\alias{constructTeststuitesTag}
\title{Construct the JUnit testsuites tag}
\usage{
constructTeststuitesTag(testResults)
}
\arguments{
\item{testResults}{\link[tinytest]{tinytests}-object to convert into a JUnit xml object.
Usually the result of calling \code{\link[tinytest:test_package]{tinytest::test_package()}}.}
}
\value{
\code{XMLtag}: with tag-name = "testsuites". This is the root of the JUnit xml document.
}
\description{
Convert the tinytests object containing test across possibly multiple files into a JUnit
testsuites tag.
}
\details{
Reference for JUnit XML format: https://llg.cubic.org/docs/junit/
}
\author{
ltuijnder
}
