% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructJUnitTag.R
\name{constructTeststuiteTag}
\alias{constructTeststuiteTag}
\title{Construct JUnit testsuite tag}
\usage{
constructTeststuiteTag(testResultsSingleFile)
}
\arguments{
\item{testResultsSingleFile}{\code{tinytests} with all test results from that ran in the specified file.}
}
\value{
\code{XMLtag}: with tag-name = "testsuite" and contains all the tests results of the file.
}
\description{
Given all the tinytests result from a single file. Construct its the "testsuite" of t
}
\author{
ltuijnder
}
