% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructJUnitTag.R
\name{constructTestsuiteTag}
\alias{constructTestsuiteTag}
\title{Construct JUnit \verb{</testsuite>} tag}
\usage{
constructTestsuiteTag(testResultsSingleFile)
}
\arguments{
\item{testResultsSingleFile}{\code{tinytesta}-object with all test results of a specified test file.}
}
\value{
\code{XMLtag}: with tag-name = \verb{</testsuite>} that contains all the test results per test file.
}
\description{
Construct the \verb{</testsuite>} tag of a \code{tinytest}, given all the \code{tinytest} results from a single test file.
}
