% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_2_junit.R
\name{convert_2_junit}
\alias{convert_2_junit}
\title{Convert tinytest output to JUnit xml}
\usage{
convert_2_junit(df)
}
\arguments{
\item{df}{output}
}
\value{
\code{character(n)} a character vector where each line represents a line in the .xml output.
}
\description{
Convert the tinytest output dataframe to JUnit xml.
}
\details{
Reference for JUnit XML format: https://llg.cubic.org/docs/junit/
}
\author{
ltuijnder
}
