% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructJUnitTag.R
\name{errorTestcaseTag}
\alias{errorTestcaseTag}
\title{Construct a testcase-tag for an error test.}
\usage{
errorTestcaseTag(tinytest)
}
\arguments{
\item{tinytest}{a \code{tinytest} object already validated to be a "ERROR" test.}
}
\value{
a \code{testase} \code{XMLtag}
}
\description{
Construct a testcase-tag for an error test.
}
