% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{escapeXml}
\alias{escapeXml}
\title{Escape xml}
\usage{
escapeXml(x)
}
\arguments{
\item{x}{a \code{character} vector meant to be xml}
}
\value{
The same \code{character} vector x but xml escaped.
}
\description{
Escape the characters \code{&},\verb{"},\verb{'},\code{<},\code{>}
}
\seealso{
https://stackoverflow.com/a/1091953/10415129
}
