% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructJUnitTag.R
\name{failureTestcaseTag}
\alias{failureTestcaseTag}
\title{Construct a testcase-tag for a failed test.}
\usage{
failureTestcaseTag(tinytest)
}
\arguments{
\item{tinytest}{a \code{tinytest} object already validated to be a "FAILURE" test.}
}
\value{
a \code{testase} \code{XMLtag}
}
\description{
Construct a testcase-tag for a failed test.
}
