% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{format.XMLtag}
\alias{format.XMLtag}
\title{Format method for XMLtag class}
\usage{
\method{format}{XMLtag}(x, level = 0, ...)
}
\arguments{
\item{x}{an \code{XMLtag}-object}

\item{level}{print depth level. For each level 2 spaces are added to the left. The content of a
tag is automatically indented with 1 level.}

\item{...}{to ignore}
}
\value{
\code{character(1)} vector of the formatted XML tag.
}
\description{
Format S3 method for the \code{XMLtag}-class
}
\author{
Lennart Tuijnder
}
