% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isSingleLengthCharNonNA}
\alias{isSingleLengthCharNonNA}
\title{Test if single length character non NA.}
\usage{
isSingleLengthCharNonNA(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
\code{logical(1)}
}
\description{
Test if single length character non NA.
}
\author{
ltuijnder
}
