% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructJUnitTag.R
\name{nameTestcase}
\alias{nameTestcase}
\title{Helper function to construct the name of a testcase}
\usage{
nameTestcase(tinytest)
}
\arguments{
\item{tinytest}{a \code{tinytest} object. (does not matter what result)}
}
\value{
\code{character(1)} the testcase name to use for this tinytest object.
}
\description{
Helper function to construct the name of a testcase. Note, the charater is already xml escaped.
}
