% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructJUnitTag.R
\name{passedTestcaseTag}
\alias{passedTestcaseTag}
\title{Construct a testcase-tag for a passed test.}
\usage{
passedTestcaseTag(tinytest)
}
\arguments{
\item{tinytest}{a \code{tinytest} object already validated to be a "PASSED" test.}
}
\value{
a \code{testase} \code{XMLtag}
}
\description{
Construct a testcase-tag for a passed test.
}
