% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{print.XMLtag}
\alias{print.XMLtag}
\title{Print method for XMLtag class.}
\usage{
\method{print}{XMLtag}(x, ...)
}
\arguments{
\item{x}{a \code{XMLtag}-object}

\item{...}{to be ignored}
}
\value{
\code{invisibly} the string that was printed to stdout.
}
\description{
Print method for XMLtag class.
}
