% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_and_convert.R
\name{run_and_convert}
\alias{run_and_convert}
\title{tinytest the specified directory and save the result to JUnit}
\usage{
run_and_convert(dir = ".", file = NULL)
}
\arguments{
\item{file}{\code{character(1) | NULL} .xml file path to save the JUnit tests results to. Or NULL to print
the xml to stdout. By default NULL.}

\item{path}{\code{character(1)} file path to root of the package. Either a source root repositry or a
built package with test = TRUE.}
}
\value{
to be ignored.
}
\description{
tinytest a specified with tinytest and save the results in JUnit xml format to the specified file.
Or print in to stdout.
}
\author{
ltuijnder
}
