% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructJUnitTag.R
\name{sideeffectTestcaseTag}
\alias{sideeffectTestcaseTag}
\title{Construct a testcase-tag for a side-effect test.}
\usage{
sideeffectTestcaseTag(tinytest)
}
\arguments{
\item{tinytest}{a \code{tinytest} object already validated to be a "SIDE-EFFECT" test.}
}
\value{
a \code{testase} \code{XMLtag}
}
\description{
Construct a testcase-tag for a side-effect test.
}
