% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTestDir.R
\name{[.tinytests2JUnit}
\alias{[.tinytests2JUnit}
\alias{tinytests2JUnit}
\title{tinytestJUnit test results}
\usage{
\method{[}{tinytests2JUnit}(x, i)
}
\arguments{
\item{x}{\code{tinytests2JUnit} object to subset.}

\item{i}{object to subset the \code{x} with.}
}
\description{
An object of class \code{tinytests2JUnit}. Note the plurar. A subclass of \code{\link[tinytest:tinytests]{tinytest::tinytests()}}
containing extra info recordings that are used in the export to JUnit.
}
\details{
Following details are recorded when running the tests files and stored as additional attributes
to the object:
\itemize{
\item \strong{fileDurations}: \code{named-numeric(n)}. Names = filename of tests files, value = duration
in seconds on how long the test file took to run.
\item \strong{fileTimestamps}: \code{named-character(n)}. Names = filename of tests files, value = timestamp
when the test was invoked.
\item \strong{fileHostnames}: \code{named-character(n)}. Names = filename of tests files, value = The hostname
of the system that ran the tests. (Usefull in combination with clusters).
\item \strong{disabled}: \code{character}. A character vector of filenames where no tests were ran.
They are flagged as disabled tests.
}
}
