% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructJUnitTag.R
\name{constructTestcaseTag}
\alias{constructTestcaseTag}
\title{Construct JUnit \verb{</testcase>} tag}
\usage{
constructTestcaseTag(tinytest)
}
\arguments{
\item{`tinytest`}{a \code{tinytests}-object representing an individual test case.}
}
\value{
\code{XMLtag}: with tag-name = \code{tinytest} and contains the test result per test.
}
\description{
Construct JUnit \verb{</testcase>} tag based on a single \code{tinytest} result.
}
