% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructJUnitTag.R
\name{constructTestsuitesTag}
\alias{constructTestsuitesTag}
\title{Construct the JUnit \verb{</testsuites>} tag}
\usage{
constructTestsuitesTag(testResults)
}
\arguments{
\item{testResults}{\verb{tinytests2Junit | }tinytests\verb{object to convert into a JUnit XML object  Usually the result of [runTestDir()] for a}tinytests2JUnit\verb{object or native}tinytests`
produced by \code{\link[tinytest:test_package]{tinytest::test_package()}} or \code{\link[tinytest:run_test_dir]{tinytest::run_test_dir()}}.}
}
\value{
\code{XMLtag}: with tag-name = \verb{</testsuites>}. This is the root of the JUnit XML document.
}
\description{
Convert the \code{tinytests2Junit} or \code{tinytests}-object containing test across
possibly multiple files into a JUnit \verb{</testsuites>} tag. More details are reported to the
JUnit if a \code{tinytests2JUnit} object compared to the native \code{tinytests} object.
}
\details{
Reference for JUnit XML format: https://llg.cubic.org/docs/junit/

See details \code{\link[=runTestDir]{runTestDir()}} which additional info is recorded.
}
