% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructJUnitTag.R
\name{constructTestsuiteTag}
\alias{constructTestsuiteTag}
\title{Construct JUnit \verb{</testsuite>} tag}
\usage{
constructTestsuiteTag(testsFile, id)
}
\arguments{
\item{testsFile}{\code{tinytests2JUnit | tinytests} -object with all test
results of a specified test file. At least a single test is expected.}

\item{id}{\code{integer(1)} testsuite id.}
}
\value{
\code{XMLtag}: with tag-name = \verb{</testsuite>}
that contains all the test results per test file.
}
\description{
Construct the \verb{</testsuite>} tag of a \code{tinytest}, given all the \code{tinytest} results
from a single test file.
}
\details{
In case a \code{tinytest2JUnit} is provided following additional info can be reported:
\itemize{
\item testsuite duration.
\item timestamp when the testsuite was performed.
\item hostname where the testsuite was ran.
}
}
