% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{tag}
\alias{tag}
\title{XML tag}
\usage{
tag(name, ...)
}
\arguments{
\item{name}{\code{character(1)} specifying the name of the tag.}

\item{...}{named arguments represent xml attributes, while unnamed arguments represents tag content.}
}
\value{
a \code{XMLtag}-object. The name of element is the tag-name.
The value of the element represents the tag content. The attributes of the tag are attached to
the tag_content (to not conflict with the names attribute of the tag)
}
\description{
Create a list object that roughly mimics the behaviour of a simplistic xml tag element.
Supported is xml tag-name, tag-attributes and tag-content.
}
\details{
The \code{xml-tag} object is a under-the-hood a list object with 2 attributes. "tag-name" and "class". The
latter of couse being "xml-tag" while the "name" attribute contains the tag name.
}
\author{
ltuijnder
}
