% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeJUnit.R
\name{writeJUnit}
\alias{writeJUnit}
\title{Write the results in a \code{tinytests}-object into JUnit xml report.}
\usage{
writeJUnit(tinytests, file, overwrite = TRUE)
}
\arguments{
\item{tinytests}{\code{tinytests} object to convert to JUnit xml.}

\item{file}{\code{character(1)}: Full file path to the .xml file to write the JUnit xml into.
Example: "/home/ltuijnder/documents/results.xml".}

\item{overwrite}{\code{logical(1)}: should the file be overwritten if it already exist?
By default TRUE.}
}
\value{
\code{invisible: TRUE}.
}
\description{
Write the \code{\link[tinytest:tinytests]{tinytest::tinytests()}}-object to a JUnit XML reporting file.
}
\section{Errors}{

In case of overwrite = FALSE and the file already exist an error is thrown.
}

\seealso{
The JUnit-xml report format: https://llg.cubic.org/docs/junit/
}
\author{
ltuijnder
}
